wip: migrate cwlog
